pyproject.toml
[project]
name = "book-llm-novv1"
version = "0.1.0"
description = "BOOK LLM novv1 - small local book Q&A + summarizer (RAG)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pypdf>=4.2.0",
  "sentence-transformers>=3.0.1",
  "numpy>=1.26.4",
  "llama-cpp-python>=0.2.86",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.1",
]

[project.scripts]
book-llm = "book_llm.cli:app"

[tool.ruff]
line-length = 100
